Add scaffolding to support new subsampling method

+++ src/plastimatch/register/plm_stages.cxx
@@ -271,12 +271,13 @@ do_registration_stage (
 }
 
 static void
-set_auto_subsampling (int subsample_rate[], Plm_image *pli)
+set_auto_subsampling (float subsample_rate[], Plm_image *pli)
 {
     Plm_image_header pih (pli);
-    
+
+    /* GCS LEFT OFF HERE */
     for (int d = 0; d < 3; d++) {
-        subsample_rate[d] = (pih.Size(d)+99) / 100;
+        subsample_rate[d] = (float) ((pih.Size(d)+99) / 100);
     }
 }
